# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/music/search": {
      "get": {
        "operationId": "MusicController_search",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "minLength": 1,
              "maxLength": 100,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrackSummaryDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired access token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Search tracks (Deezer). Results never include preview links",
        "tags": [
          "Music (Deezer)"
        ]
      }
    },
    "/music/tracks/{providerTrackId}/preview": {
      "get": {
        "operationId": "MusicController_preview",
        "parameters": [
          {
            "name": "providerTrackId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackPreviewDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired access token"
          },
          "404": {
            "description": "Unknown track, or no preview"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "A fresh 30 s preview link (expires in ~15 min: fetch it right before playing)",
        "tags": [
          "Music (Deezer)"
        ]
      }
    },
...
    "/events/{id}": {
      "get": {
        "operationId": "EventsController_view",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lat",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "lng",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventViewDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired access token"
          },
          "404": {
            "description": "Event not found — also answered for private events you are not in, so their existence is not revealed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Event details, your role, and whether you can vote (send ?lat=&lng= for geo events)",
        "tags": [
          "Events (Track Vote)"
        ]
      },
      "patch": {
        "operationId": "EventsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventViewDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Missing, invalid or expired access token"
          },
          "403": {
            "description": "Only the event owner can do this"
          },
          "404": {
            "description": "Event not found — also answered for private events you are not in, so their existence is not revealed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Edit an event (owner). Switching license to geo requires all geo fields",
        "tags": [
          "Events (Track Vote)"
        ]
      },
      "delete": {
        "operationId": "EventsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "Missing, invalid or expired access token"
          },
          "403": {
            "description": "Only the event owner can do this"
          },
          "404": {
            "description": "Event not found — also answered for private events you are not in, so their existence is not revealed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete an event and its queue (owner)",
        "tags": [
          "Events (Track Vote)"
        ]
      }
    },
...
    "/events/{id}/queue": {
      "get": {
        "operationId": "QueueController_view",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueViewDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired access token"
          },
          "404": {
            "description": "Event not found — also answered for private events you are not in, so their existence is not revealed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Now playing + the ranked queue (score DESC, earliest suggestion first), with votedByMe",
        "tags": [
          "Queue (Track Vote)"
        ]
      }
    },
    "/events/{id}/tracks": {
      "post": {
        "operationId": "QueueController_suggest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuggestTrackDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueTrackDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired access token"
          },
          "403": {
            "description": "Not allowed to vote/suggest. Body has a machine-readable `reason`: not_invited | not_started | ended | location_required | outside_area"
          },
          "404": {
            "description": "Event or Deezer track not found"
          },
          "409": {
            "description": "This track is already in the queue"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Suggest a track by its Deezer id (the server fetches the details itself)",
        "tags": [
          "Queue (Track Vote)"
        ]
      }
    },
    "/events/{id}/tracks/{trackId}/vote": {
      "post": {
        "operationId": "QueueController_vote",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trackId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoteResultDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired access token"
          },
          "403": {
            "description": "Not allowed to vote/suggest. Body has a machine-readable `reason`: not_invited | not_started | ended | location_required | outside_area"
          },
          "404": {
            "description": "Event or track not found"
          },
          "409": {
            "description": "The track is no longer queued (playing or played)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Vote for a queued track (idempotent: voting twice changes nothing)",
        "tags": [
          "Queue (Track Vote)"
        ]
      },
      "delete": {
        "operationId": "QueueController_unvote",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trackId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lat",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "lng",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoteResultDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired access token"
          },
          "403": {
            "description": "Not allowed to vote/suggest. Body has a machine-readable `reason`: not_invited | not_started | ended | location_required | outside_area"
          },
          "404": {
            "description": "Event or track not found"
          },
          "409": {
            "description": "The track is no longer queued"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove your vote (idempotent). Geo events: location as ?lat=&lng=",
        "tags": [
          "Queue (Track Vote)"
        ]
      }
    },
    "/events/{id}/next": {
      "post": {
        "operationId": "QueueController_next",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NextTrackDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueBroadcastDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired access token"
          },
          "403": {
            "description": "Only the event owner controls playback"
          },
          "404": {
            "description": "Event not found — also answered for private events you are not in, so their existence is not revealed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Owner: the playing track becomes played, the top of the queue starts playing. A stale currentTrackId does nothing",
        "tags": [
          "Queue (Track Vote)"
        ]
      }
    }
  },
  "info": {
    "title": "Music Room API",
    "description": "REST API of Music Room (Track Vote events, accounts, music search).\n\n**Auth:** log in with `POST /auth/login` (or a social login), then click **Authorize** and paste the `accessToken`.\nAccess tokens last 15 minutes; get a new pair with `POST /auth/refresh`.\n\n**Real-time** (Socket.IO namespace `/events`): see `docs/realtime.md` in the repository.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "The accessToken from /auth/login"
      }
    },
    "schemas": {
...
      "TrackSummaryDto": {
        "type": "object",
        "properties": {
          "provider": {
            "enum": [
              "deezer"
            ],
            "type": "string"
          },
          "coverUrl": {
            "type": "string",
            "nullable": true
          },
          "isrc": {
            "type": "string",
            "nullable": true,
            "description": "International Standard Recording Code (same recording across providers).",
            "example": "GBDUW0000053"
          },
          "providerTrackId": {
            "type": "string",
            "description": "The provider's track id (Deezer: digits).",
            "example": "3135553"
          },
          "title": {
            "type": "string",
            "example": "One More Time"
          },
          "artist": {
            "type": "string",
            "example": "Daft Punk"
          },
          "album": {
            "type": "string",
            "example": "Discovery"
          },
          "durationSec": {
            "type": "number",
            "example": 320
          }
        },
        "required": [
          "provider",
          "coverUrl",
          "isrc",
          "providerTrackId",
          "title",
          "artist",
          "album",
          "durationSec"
        ]
      },
      "TrackPreviewDto": {
        "type": "object",
        "properties": {
          "expiresAt": {
            "type": "string",
            "description": "When the link stops working (≈ 15 minutes).",
            "format": "date-time"
          },
          "url": {
            "type": "string",
            "description": "Signed MP3 link (30 s). Never store it: it expires."
          }
        },
        "required": [
          "expiresAt",
          "url"
        ]
      },
...
      "EventGeoDto": {
        "type": "object",
        "properties": {
          "startsAt": {
            "type": "string",
            "format": "date-time"
          },
          "endsAt": {
            "type": "string",
            "format": "date-time"
          },
          "lat": {
            "type": "number",
            "example": 48.8966
          },
          "lng": {
            "type": "number",
            "example": 2.3185
          },
          "radiusM": {
            "type": "number",
            "example": 200
          }
        },
        "required": [
          "startsAt",
          "endsAt",
          "lat",
          "lng",
          "radiusM"
        ]
      },
      "UserRefDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "displayName"
        ]
      },
      "ParticipationDto": {
        "type": "object",
        "properties": {
          "reason": {
            "description": "Present when `allowed` is false.",
            "enum": [
              "not_member",
              "not_invited",
              "not_started",
              "ended",
              "location_required",
              "outside_area"
            ],
            "type": "string"
          },
          "allowed": {
            "type": "boolean",
            "description": "Whether you can vote and suggest tracks right now."
          }
        },
        "required": [
          "allowed"
        ]
      },
      "EventViewDto": {
        "type": "object",
        "properties": {
          "visibility": {
            "enum": [
              "public",
              "private"
            ],
            "type": "string"
          },
          "license": {
            "description": "open = anyone who can see it · invited = owner + invited accounts · geo = inside an area during a time window",
            "enum": [
              "open",
              "invited",
              "geo"
            ],
            "type": "string"
          },
          "geo": {
            "nullable": true,
            "description": "Only when license is 'geo'.",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/EventGeoDto"
              }
            ]
          },
          "owner": {
            "$ref": "#/components/schemas/UserRefDto"
          },
          "myRole": {
            "nullable": true,
            "description": "Your membership; null = not a member (only possible for public events).",
            "enum": [
              "owner",
              "invited",
              "guest"
            ],
            "type": "string"
          },
          "inviteCode": {
            "type": "string",
            "nullable": true,
            "description": "8 characters; only shown to members.",
            "example": "KVAMDDE4"
          },
          "participation": {
            "description": "For license 'geo' this depends on the location you sent (?lat=&lng=).",
            "allOf": [
              {
                "$ref": "#/components/schemas/ParticipationDto"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "example": "Friday Party"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "visibility",
          "license",
          "geo",
          "owner",
          "myRole",
          "inviteCode",
          "participation",
          "createdAt",
          "id",
          "name",
          "description"
        ]
      },
...
      "UpdateEventDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "maxLength": 500
          },
          "visibility": {
            "type": "string",
            "enum": [
              "public",
              "private"
            ]
          },
          "license": {
            "type": "string",
            "enum": [
              "open",
              "invited",
              "geo"
            ]
          },
          "geoLat": {
            "type": "number"
          },
          "geoLng": {
            "type": "number"
          },
          "geoRadiusM": {
            "type": "number",
            "minimum": 10,
            "maximum": 50000
          },
          "startsAt": {
            "type": "string"
          },
          "endsAt": {
            "type": "string"
          }
        }
      },
...
      "QueueTrackDto": {
        "type": "object",
        "properties": {
          "provider": {
            "enum": [
              "deezer"
            ],
            "type": "string"
          },
          "coverUrl": {
            "type": "string",
            "nullable": true
          },
          "isrc": {
            "type": "string",
            "nullable": true,
            "description": "International Standard Recording Code (same recording across providers).",
            "example": "GBDUW0000053"
          },
          "status": {
            "enum": [
              "queued",
              "playing",
              "played"
            ],
            "type": "string"
          },
          "suggestedBy": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRefDto"
              }
            ]
          },
          "suggestedAt": {
            "type": "string",
            "description": "Tie-break: at equal score the earliest suggestion plays first.",
            "format": "date-time"
          },
          "votedByMe": {
            "type": "boolean",
            "description": "Whether you voted for this track."
          },
          "id": {
            "type": "string",
            "description": "The queue entry's id — use it to vote (not the provider's track id)."
          },
          "score": {
            "type": "number",
            "description": "Number of votes."
          },
          "providerTrackId": {
            "type": "string",
            "description": "The provider's track id (Deezer: digits).",
            "example": "3135553"
          },
          "title": {
            "type": "string",
            "example": "One More Time"
          },
          "artist": {
            "type": "string",
            "example": "Daft Punk"
          },
          "album": {
            "type": "string",
            "example": "Discovery"
          },
          "durationSec": {
            "type": "number",
            "example": 320
          }
        },
        "required": [
          "provider",
          "coverUrl",
          "isrc",
          "status",
          "suggestedBy",
          "suggestedAt",
          "votedByMe",
          "id",
          "score",
          "providerTrackId",
          "title",
          "artist",
          "album",
          "durationSec"
        ]
      },
      "QueueViewDto": {
        "type": "object",
        "properties": {
          "nowPlaying": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/QueueTrackDto"
              }
            ]
          },
          "upcoming": {
            "description": "Ranked: score DESC, then suggestedAt ASC.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QueueTrackDto"
            }
          }
        },
        "required": [
          "nowPlaying",
          "upcoming"
        ]
      },
      "SuggestTrackDto": {
        "type": "object",
        "properties": {
          "providerTrackId": {
            "type": "string",
            "pattern": "^\\d{1,20}$"
          },
          "lat": {
            "type": "number"
          },
          "lng": {
            "type": "number"
          }
        },
        "required": [
          "providerTrackId"
        ]
      },
      "LocationDto": {
        "type": "object",
        "properties": {
          "lat": {
            "type": "number"
          },
          "lng": {
            "type": "number"
          }
        }
      },
      "VoteResultDto": {
        "type": "object",
        "properties": {
          "trackId": {
            "type": "string"
          },
          "score": {
            "type": "number",
            "description": "The track's score after your request."
          },
          "votedByMe": {
            "type": "boolean"
          }
        },
        "required": [
          "trackId",
          "score",
          "votedByMe"
        ]
      },
      "NextTrackDto": {
        "type": "object",
        "properties": {
          "currentTrackId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          }
        }
      },
      "BroadcastTrackDto": {
        "type": "object",
        "properties": {
          "provider": {
            "enum": [
              "deezer"
            ],
            "type": "string"
          },
          "coverUrl": {
            "type": "string",
            "nullable": true
          },
          "isrc": {
            "type": "string",
            "nullable": true,
            "description": "International Standard Recording Code (same recording across providers).",
            "example": "GBDUW0000053"
          },
          "status": {
            "enum": [
              "queued",
              "playing",
              "played"
            ],
            "type": "string"
          },
          "suggestedBy": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRefDto"
              }
            ]
          },
          "suggestedAt": {
            "type": "string",
            "description": "Tie-break: at equal score the earliest suggestion plays first.",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "description": "The queue entry's id — use it to vote (not the provider's track id)."
          },
          "score": {
            "type": "number",
            "description": "Number of votes."
          },
          "providerTrackId": {
            "type": "string",
            "description": "The provider's track id (Deezer: digits).",
            "example": "3135553"
          },
          "title": {
            "type": "string",
            "example": "One More Time"
          },
          "artist": {
            "type": "string",
            "example": "Daft Punk"
          },
          "album": {
            "type": "string",
            "example": "Discovery"
          },
          "durationSec": {
            "type": "number",
            "example": 320
          }
        },
        "required": [
          "provider",
          "coverUrl",
          "isrc",
          "status",
          "suggestedBy",
          "suggestedAt",
          "id",
          "score",
          "providerTrackId",
          "title",
          "artist",
          "album",
          "durationSec"
        ]
      },
      "QueueBroadcastDto": {
        "type": "object",
        "properties": {
          "nowPlaying": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/BroadcastTrackDto"
              }
            ]
          },
          "upcoming": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BroadcastTrackDto"
            }
          },
          "eventId": {
            "type": "string"
          }
        },
        "required": [
          "nowPlaying",
          "upcoming",
          "eventId"
        ]
      }
    }
  }
}